dockerization/app/requirement.txt
Flask==2.3.2
mysql-connector-python==9.2.0
